// src/types/Book.ts
export interface Book {
  key: string;
  title: string;
  author_name?: string[];
  author_key?: string[];
  cover_i?: number;
  first_publish_year?: number;
  author_alternative_name?: string[];
  isbn?: string[];
  contributor?: string[];
  cover_edition_key?: string;
  ddc?: string[];
  ebook_access?: string;
  ebook_count_i?: number;
  edition_count?: number;
  edition_key?: string[];
  first_sentence?: string[];
  format?: string[];
  has_fulltext?: boolean;
  ia_collection?: string[];
  ia_collection_s?: string;
  language?: string[];
  last_modified_i?: number;
  number_of_pages_median?: number;
  osp_count?: number;
  public_scan_b?: boolean;
  publish_date?: string[];
  publish_place?: string[];
  publish_year?: number[];
  publisher?: string[];
  title_suggest?: string;
  title_sort?: string;
  type?: string;
  subject?: string[];
  place?: string[];
  time?: string[];
  person?: string[];
  ratings_average?: number;
  ratings_sortable?: number;
  ratings_count?: number;
  ratings_count_1?: number;
  ratings_count_2?: number;
  ratings_count_3?: number;
  ratings_count_4?: number;
  ratings_count_5?: number;
  readinglog_count?: number;
  want_to_read_count?: number;
  currently_reading_count?: number;
  already_read_count?: number;
  publisher_facet?: string[];
  person_key?: string[];
  time_facet?: string[];
  place_key?: string[];
  person_facet?: string[];
  subject_facet?: string[];
  subject_key?: string[];
  ddc_sort?: string;
  time_key?: string[];
}